clean up of rads, more error handling, add key info to eb requests
