+++ rads/setup.cfg
@@ -14,12 +14,8 @@ maintainer = Ryan Medick
 maintainer_email = rmedick@pdx.edu
 url = https://github.com/oresat/uniclogs-software
 classifiers =
-    Development Status :: 5 - Production/Stable
-    Environment :: Console
-    Intended Audience :: Developers
     License :: OSI Approved :: GPL v3 License
     Programming Language :: Python
-    Topic :: Software Development :: Libraries :: Python Modules
 
 [options]
 package_dir=
@@ -33,4 +29,4 @@ where = src
 
 [options.entry_points]
 console_scripts =
-    rads = rads.ui.director:main
+    rads = rads.ui.main_menu:main
