+++ rads/src/rads/ui/tables/request_table.py
@@ -1,6 +1,12 @@
-from rads.database.db_interface import query_upcomming_requests,\
-        update_approve_deny
+"""
+The classes to handle display eb passes.
+"""
+
+from rads.database.query import query_upcomming_requests
+from rads.database.update import update_approve_deny
 from rads.database.request_data import RequestData
+from rads.command.schedule_pass import Schedule_Pass
+from loguru import logger
 import reverse_geocoder as rg
 
 _DT_STR_FORMAT = "%Y/%m/%d %H:%M:%S"
@@ -8,23 +14,27 @@ _STR_FORMAT = "{:7} | {:8} | {:15} | {:^5} | {:19} | {:19} | {:30}"
 
 
 class Request(RequestData):
+    """
+    A wrapper class ontop of request data for the ui to use.
+    """
     def __init__(self, request):
         super().__init__(
-                request.id,
-                request.user_token,
-                request.pass_id,
-                request.is_approved,
-                request.is_sent,
-                request.created_dt,
-                request.updated_dt,
-                request.observation_type,
-                request.pass_data.gs_latitude_deg,
-                request.pass_data.gs_longitude_deg,
-                request.pass_data.gs_elevation_m,
-                request.pass_data.aos_utc,
-                request.pass_data.los_utc
-                )
+            request.id,
+            request.user_token,
+            request.pass_id,
+            request.is_approved,
+            request.is_sent,
+            request.created_dt,
+            request.updated_dt,
+            request.observation_type,
+            request.pass_data.gs_latitude_deg,
+            request.pass_data.gs_longitude_deg,
+            request.pass_data.gs_elevation_m,
+            request.pass_data.aos_utc,
+            request.pass_data.los_utc
+            )
         self.deny_count = 0
+        self.geo = request.geo
 
     def __str__(self):
 
@@ -50,20 +60,26 @@ class Request(RequestData):
             loc = " "
 
         return _STR_FORMAT.format(
-                self.id,
-                ad_status,
-                obs_type,
-                sent_status,
-                self.pass_data.aos_utc.strftime(_DT_STR_FORMAT),
-                self.pass_data.los_utc.strftime(_DT_STR_FORMAT),
-                loc
-                )
+            self.id,
+            ad_status,
+            obs_type,
+            sent_status,
+            self.pass_data.aos_utc.strftime(_DT_STR_FORMAT),
+            self.pass_data.los_utc.strftime(_DT_STR_FORMAT),
+            loc
+            )
 
     def deny(self):
+        """
+        Deny request.
+        """
         self.deny_count += 1
         self.is_approved = False
 
     def undeny(self):
+        """
+        Undo the denied request.
+        """
         if self.deny_count > 0:
             self.deny_count -= 1
             if self.deny_count == 0:
@@ -71,6 +87,10 @@ class Request(RequestData):
 
 
 class RequestTable():
+    """
+    A list of request that approved and upcomming.
+    """
+
     def __init__(self):
         self.data = []
         coordinates = []
@@ -78,7 +98,7 @@ class RequestTable():
         requests = query_upcomming_requests()
         for r in requests:
             self.data.append(Request(r))
-            coor = (r.pass_data.gs_latitude_deg,  r.pass_data.gs_longitude_deg)
+            coor = (r.pass_data.gs_latitude_deg, r.pass_data.gs_longitude_deg)
             coordinates.append(coor)
 
         locations = rg.search(coordinates, verbose=False)
@@ -88,14 +108,14 @@ class RequestTable():
 
         self.data_len = len(self.data)
         self.header = _STR_FORMAT.format(
-                "ID",
-                "Status",
-                "Type",
-                "Sent",
-                "AOS",
-                "LOS",
-                "City, State"
-                )
+            "ID",
+            "Status",
+            "Type",
+            "Sent",
+            "AOS",
+            "LOS",
+            "City, State"
+            )
 
     def __getitem__(self, index):
         return self.data[index]
@@ -104,4 +124,12 @@ class RequestTable():
         return self.data_len
 
     def save(self):
+        """
+        Save data to db and update cosmos.
+        """
         update_approve_deny(self.data)
+        try:
+            schedule_pass = Schedule_Pass(self.data)
+            schedule_pass.schedule_all()
+        except Exception as e:
+            logger.error("cosmos command interface failed with: {}".format(e))
